fixed bug in songMenuButton, song list screen and song detail page refactored
