+++ sputofy_2/lib/screens/SongDetailScreen/components/seek_bar.dart
@@ -1,0 +1,63 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:sputofy_2/theme/palette.dart';
+
+class SeekBar extends StatefulWidget {
+  final Duration duration;
+  final Duration position;
+  final ValueChanged<Duration>? onChanged;
+  final ValueChanged<Duration>? onChangeEnd;
+
+  SeekBar({
+    required this.duration,
+    required this.position,
+    this.onChanged,
+    this.onChangeEnd,
+  });
+
+  @override
+  _SeekBarState createState() => _SeekBarState();
+}
+
+class _SeekBarState extends State<SeekBar> {
+  double? _dragValue;
+  bool _dragging = false;
+
+  @override
+  Widget build(BuildContext context) {
+    final value = min(_dragValue ?? widget.position.inMilliseconds.toDouble(),
+        widget.duration.inMilliseconds.toDouble());
+    if (_dragValue != null && !_dragging) {
+      _dragValue = null;
+    }
+    return Stack(
+      children: [
+        Slider(
+          activeColor: kAccentColor,
+          inactiveColor: kPrimaryColor,
+          min: 0.0,
+          max: widget.duration.inMilliseconds.toDouble(),
+          value: value,
+          onChanged: (value) {
+            if (!_dragging) {
+              _dragging = true;
+            }
+            setState(() {
+              _dragValue = value;
+            });
+            if (widget.onChanged != null) {
+              widget.onChanged!(Duration(milliseconds: value.round()));
+            }
+          },
+          onChangeEnd: (value) {
+            if (widget.onChangeEnd != null) {
+              widget.onChangeEnd!(Duration(milliseconds: value.round()));
+            }
+            _dragging = false;
+          },
+        ),
+      ],
+    );
+  }
+}
